[fix] gws elasticsearch setting discussion
